requirements.txt
streamlit==1.22.0
pandas==2.0.3
python-dotenv==1.0.0
crewai==0.1.0
pdfplumber==0.8.0
PyPDF2==3.0.1
openai
langchain
docx

# All dependencies using python 3.11.8
# streamlit==1.22.0
# pandas==2.0.3
# python-dotenv==1.0.0
# crewai==0.1.0
# pdfplumber==0.8.0
# PyPDF2==3.0.1
# openai==0.28.1
# langchain==0.0.335
# altair<5,>=3.2.0
# blinker>=1.0.0
# cachetools>=4.0
# click>=7.0
# importlib-metadata>=1.4
# numpy==1.26.4
# packaging>=14.1
# pillow>=6.2.0
# protobuf<4,>=3.12
# pyarrow>=4.0
# pympler>=0.9
# python-dateutil==2.9.0.post0
# requests>=2.4
# rich>=10.11.0
# tenacity<9,>=8.0.0
# toml
# typing-extensions>=3.10.0.0
# tzlocal>=1.1
# validators>=0.2
# gitpython!=3.1.19
# pydeck>=0.1.dev5
# tornado>=6.0.3
# watchdog==4.0.2
# pytz>=2020.1
# tzdata>=2022.1
# pydantic<3.0.0,>=2.4.2
# pdfminer.six==20221105
# Wand>=0.6.10
# charset-normalizer>=2.0.0
# cryptography>=36.0.0
# tqdm==4.66.5
# aiohttp==3.10.3
# PyYAML>=5.3
# SQLAlchemy<3,>=1.4
# anyio<4.0
# dataclasses-json<0.7,>=0.5.7
# jsonpatch<2.0,>=1.33
# langsmith<0.1.0,>=0.0.63
# aiohappyeyeballs>=2.3.0
# aiosignal>=1.1.2
# attrs>=17.3.0
# frozenlist>=1.1.1
# multidict<7.0,>=4.5
# yarl<2.0,>=1.0
# entrypoints==0.4
# jinja2==3.1.4
# jsonschema>=3.0
# toolz==0.12.1
# idna>=2.8
# sniffio>=1.1
# colorama==0.4.6
# marshmallow<4.0.0,>=3.18.0
# typing-inspect<1,>=0.4.0
# gitdb<5,>=4.0.1
# zipp>=0.5
# jsonpointer>=1.9
# annotated-types>=0.4.0
# pydantic-core==2.20.1
# pywin32>=226
# six>=1.5
# urllib3<3,>=1.21.1
# certifi>=2017.4.17
# markdown-it-py>=2.2.0
# pygments<3.0.0,>=2.13.0
# greenlet!=0.4.17
# cffi>=1.12
# smmap<6,>=3.0.1
# MarkupSafe>=2.0
# jsonschema-specifications>=2023.03.6
# referencing>=0.28.4
# rpds-py>=0.7.1
# mdurl~=0.1
# mypy-extensions>=0.3.0
# pycparser==2.22
